+++ src/main/java/docAsTest/action/CompareFileAction.java
@@ -5,6 +5,7 @@ import com.intellij.diff.DiffManager;
 import com.intellij.diff.contents.DiffContent;
 import com.intellij.diff.requests.DiffRequest;
 import com.intellij.diff.requests.SimpleDiffRequest;
+import com.intellij.openapi.actionSystem.ActionUpdateThread;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.actionSystem.CommonDataKeys;
 import com.intellij.openapi.actionSystem.PlatformDataKeys;
@@ -34,6 +35,12 @@ public class CompareFileAction extends DocAsTestAction {
             return approved.isPresent() && receveived.isPresent();
         }
     }
+
+    @Override
+    public @NotNull ActionUpdateThread getActionUpdateThread() {
+        return ActionUpdateThread.BGT;
+    }
+
     @Override
     public void update(AnActionEvent e) {
         LOG.debug("CompareFileAction.update " + this.getClass().getName());